Implemented feedback from PR#209 removing utils package
