+++ pysmt/cmd/installers/msat.py
@@ -16,8 +16,7 @@ import os
 import re
 import sys
 
-from pysmt.cmd.installers.base import SolverInstaller
-from pysmt.cmd.installers.utils import TemporaryPath
+from pysmt.cmd.installers.base import SolverInstaller, TemporaryPath
 
 
 class MSatInstaller(SolverInstaller):
